+++ packages/@aws-cdk-testing/cli-integ/lib/integ-test.ts
@@ -26,6 +26,35 @@ export interface TestContext {
 }
 
 /**
+ * Retry wrapper that executes a test callback up to maxAttempts times
+ *
+ * If any attempt succeeds, it returns immediately. If all attempts fail,
+ * it throws the last error encountered.
+ */
+export function withRetry<T extends TestContext>(
+  callback: (context: T) => Promise<void>,
+  maxAttempts: number = 2,
+): (context: T) => Promise<void> {
+  return async (context: T) => {
+    let lastError;
+
+    for (let attempt = 1; attempt <= maxAttempts; attempt++) {
+      try {
+        await callback(context);
+        return;
+      } catch (error) {
+        lastError = error;
+        if (attempt < maxAttempts) {
+          context.log(`Attempt ${attempt}/${maxAttempts} failed: ${error}. Retrying...`);
+        }
+      }
+    }
+
+    throw lastError;
+  };
+}
+
+/**
  * A wrapper for jest's 'test' which takes regression-disabled tests into account and prints a banner
  */
 export function integTest(
