+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cli-lib-alpha/cdk-lib-cli-lib-list.integtest.ts
@@ -1,12 +1,12 @@
-import { integTest, withCliLibFixture } from '../../../lib';
+import { integTest, withCliLibFixture, withRetry } from '../../../lib';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
 
 integTest(
   'cli-lib list',
-  withCliLibFixture(async (fixture) => {
+  withRetry(withCliLibFixture(async (fixture) => {
     const listing = await fixture.cdk(['list'], { captureStderr: false });
     expect(listing).toContain(fixture.fullStackName('simple-1'));
-  }),
+  })),
 );
