+++ packages/@aws-cdk-testing/cli-integ/tests/cli-integ-tests/cli-lib-alpha/cdk-lib-cli-lib-synth.integtest.ts
@@ -1,10 +1,10 @@
-import { integTest, withCliLibFixture } from '../../../lib';
+import { integTest, withCliLibFixture, withRetry } from '../../../lib';
 
 jest.setTimeout(2 * 60 * 60_000); // Includes the time to acquire locks, worst-case single-threaded runtime
 
 integTest(
   'cli-lib synth',
-  withCliLibFixture(async (fixture) => {
+  withRetry(withCliLibFixture(async (fixture) => {
     await fixture.cdk(['synth', fixture.fullStackName('simple-1')]);
     expect(fixture.template('simple-1')).toEqual(
       expect.objectContaining({
@@ -22,6 +22,6 @@ integTest(
         }),
       }),
     );
-  }),
+  })),
 );
